# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flock Energy - Urja Meter Ops API",
    "description": "\nA clean, documented REST API over the legacy **Urja Meter Ops** portal.\n\nThe portal has no API. This service logs in as a normal user, manages the session,\nreverse-engineers the portal's internal endpoints, normalises its inconsistent payloads,\nand exposes the result as typed JSON that another program can consume without ever\ntouching the portal.\n\n**What it adds over the portal**\n\n* Filtering, sorting and pagination the portal cannot do (make, status, phase, build,\n  any hierarchy code).\n* The network hierarchy, reconstructed from per-meter paths into a queryable tree.\n* Real consumption, derived by differencing the portal's cumulative registers.\n* Proximity search over meter coordinates.\n* An explicit data-quality report instead of silently tidied-away inconsistencies.\n\nSee `PROTOCOL.md` for how the portal actually works, and `README.md` for design decisions\nand trade-offs.\n",
    "contact": {
      "name": "Flock Energy take-home submission"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/transformers": {
      "get": {
        "tags": [
          "Transformers"
        ],
        "summary": "List distribution transformers",
        "description": "The DT registry, enriched with a live meter count from the snapshot.\n\nThis registry is treated as the **authoritative** source for DT names: some meter records carry a stale alias (`DT-007` appears as both 'Sanganer DT 7' and 'Old Malviya Nagar Xfmr'), and the registry value wins. The conflict is still reported at `/data-quality` rather than hidden.",
        "operationId": "list_transformers_api_v1_transformers_get",
        "parameters": [
          {
            "name": "feeder_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by parent feeder, e.g. 'F-002'.",
              "title": "Feeder Code"
            },
            "description": "Filter by parent feeder, e.g. 'F-002'."
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Literal substring of code or name.",
              "title": "Search"
            },
            "description": "Literal substring of code or name."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-based page number.",
              "default": 1,
              "title": "Page"
            },
            "description": "1-based page number."
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Items per page (max 200).",
              "default": 25,
              "title": "Page Size"
            },
            "description": "Items per page (max 200)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Transformer_"
                }
              }
            }
          },
          "422": {
            "description": "Request parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The upstream portal returned an unusable response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The upstream portal is unavailable or rate limiting us.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "The upstream portal did not respond in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transformers/{code}": {
      "get": {
        "tags": [
          "Transformers"
        ],
        "summary": "Get one transformer",
        "operationId": "get_transformer_api_v1_transformers__code__get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Transformer code, e.g. 'DT-001'.",
              "title": "Code"
            },
            "description": "Transformer code, e.g. 'DT-001'."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transformer"
                }
              }
            }
          },
          "422": {
            "description": "Request parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The upstream portal returned an unusable response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The upstream portal is unavailable or rate limiting us.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "The upstream portal did not respond in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such transformer.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transformers/{code}/meters": {
      "get": {
        "tags": [
          "Transformers"
        ],
        "summary": "List meters on a transformer",
        "operationId": "get_transformer_meters_api_v1_transformers__code__meters_get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-based page number.",
              "default": 1,
              "title": "Page"
            },
            "description": "1-based page number."
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Items per page (max 200).",
              "default": 25,
              "title": "Page Size"
            },
            "description": "Items per page (max 200)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MeterSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Request parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The upstream portal returned an unusable response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The upstream portal is unavailable or rate limiting us.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "The upstream portal did not respond in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such transformer.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/data-quality": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Data-quality report",
        "description": "Every inconsistency detected in the upstream data, aggregated with counts and sample meter ids.\n\nThis exists because hiding upstream mess behind a tidy API is how downstream teams end up debugging the wrong system. It also doubles as drift detection: if the portal's data changes, these counts move.",
        "operationId": "data_quality_api_v1_data_quality_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataQualityReport"
                }
              }
            }
          },
          "422": {
            "description": "Request parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The upstream portal returned an unusable response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The upstream portal is unavailable or rate limiting us.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "The upstream portal did not respond in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataQualityIssue": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "level": {
            "type": "string",
            "title": "Level",
            "description": "'warning' or 'error'."
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "affected_count": {
            "type": "integer",
            "title": "Affected Count"
          },
          "sample_meter_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sample Meter Ids"
          }
        },
        "type": "object",
        "required": [
          "code",
          "level",
          "message",
          "affected_count"
        ],
        "title": "DataQualityIssue"
      },
      "DataQualityReport": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "meter_count": {
            "type": "integer",
            "title": "Meter Count"
          },
          "issue_count": {
            "type": "integer",
            "title": "Issue Count"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/DataQualityIssue"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "generated_at",
          "meter_count",
          "issue_count",
          "issues"
        ],
        "title": "DataQualityReport"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable slug. Branch on this, not on prose."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable, safe to display. Never a stack trace."
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Optional structured context, e.g. invalid fields or a hint."
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Correlation id; matches the X-Request-ID response header."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "The body of an :class:`ApiError` as clients receive it."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Envelope returned for **every** non-2xx response.\n\nDeclared so the OpenAPI document describes the real contract. Without it FastAPI\nadvertises its own ``HTTPValidationError`` for 422 and nothing at all for 404/5xx,\nwhich makes generated clients fail to deserialise every error they receive."
      },
      "GeoPoint": {
        "properties": {
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          }
        },
        "type": "object",
        "required": [
          "latitude",
          "longitude"
        ],
        "title": "GeoPoint"
      },
...
      "InstallStatus": {
        "type": "string",
        "enum": [
          "installed",
          "faulty",
          "decommissioned"
        ],
        "title": "InstallStatus"
      },
...
      "MeterSummary": {
        "properties": {
          "meter_id": {
            "type": "string",
            "title": "Meter Id"
          },
          "serial_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serial No"
          },
          "make": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Make"
          },
          "phase_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PhaseType"
              },
              {
                "type": "null"
              }
            ]
          },
          "install_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InstallStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "dt_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dt Code"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeoPoint"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "meter_id",
          "serial_no",
          "make",
          "phase_type",
          "install_status",
          "dt_code"
        ],
        "title": "MeterSummary",
        "description": "List-view projection. Omits the full hierarchy path to keep pages small."
      },
...
      "PageMeta": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_items": {
            "type": "integer",
            "title": "Total Items"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_previous": {
            "type": "boolean",
            "title": "Has Previous"
          }
        },
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total_items",
          "total_pages",
          "has_next",
          "has_previous"
        ],
        "title": "PageMeta"
      },
      "Page_MeterSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MeterSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "Page[MeterSummary]"
      },
      "Page_Transformer_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Transformer"
            },
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "Page[Transformer]"
      },
      "PhaseType": {
        "type": "string",
        "enum": [
          "single",
          "three"
        ],
        "title": "PhaseType"
      },
...
      "Transformer": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "feeder_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feeder Code"
          },
          "capacity_kva": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity Kva"
          },
          "meter_count": {
            "type": "integer",
            "title": "Meter Count",
            "description": "Meters attached, per the current snapshot.",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "feeder_code",
          "capacity_kva"
        ],
        "title": "Transformer",
        "description": "A distribution transformer, from the portal's authoritative DT registry."
      }
    }
  },
...
}